requirements.txt
Django==2.1.7
django-bootstrap-modal-forms==1.4.1
django-bootstrap3==11.0.0
django-pure-pagination==0.3.0
# django-crispy-forms==1.7.2
# django-filter==2.1.0
